Fix potential OOM problem
Prior to this cl, there's a potential vulnerability to
cache huge logging data in the memory.
We send the log out when user didn't trigger any event
in 1 min. However, if user keep triggering any log
event in 1 min, then there's a change to preserve a huge
number logs in the device memory. Then, it causes the OOM.

The solution is to set a logging threshold, once the current
logging number is larger than 150, then we simply send it
out immediately and clean the device memory.

Test: Log can be sent out after meeting the threshold.
Fix: 211323528
Change-Id: I9d507e8d55d6b7f7e682369edf3b5334eb2856ae

Former-commit-id: 72f823ba9e1c1be746b1d952c2f00c6dbd4cb11b