Ask user to reboot when enabling freeform and desktop mode
The introduction of FreeformTaskListener in WM shell requires a reboot
after enabling freeform support fully. Let's add a dialog explaining the
requirement and ask user to reboot the device.

Enabling desktop mode also requires a reboot to take effect, so this CL
also adds the support to it.

There is also a cancel button to dismiss the dialog without a reboot,
just in case users need to hold off the reboot for any reason.

Bug: 215437867
Test: Pressing OK reboots the device. Pressing Cancel dismisses the
dialog.
Test: atest SettingsRoboTests:FreeformWindowsPreferenceControllerTest
Test: atest SettingsRoboTests:DesktopModePreferenceControllerTest

Change-Id: I553237a29751f73afa9fd4066241a19740872f6f

Former-commit-id: 7a380f3dc6035a82a4fccb214735491049eb05ae