Resolve test failure in WorkModePreferenceControllerTest
Android T introduced a new requirement that apps registering for
non-system broadcasts via Context#registerReceiver must specify a flag
indicating whether the receiver should be exported. This flag was
added to the receiver in WorkModePreferenceController, but this broke
a test since the signature of the registerReceiver method was changed.
This commit updates the signature of the method in the test to also
expect an int parameter for the flags.

Fixes: 214906107
Test: make RunSettingsRoboTests ROBOTEST_FILTER=WorkModePreferenceControllerTest
Change-Id: I17061ab24ed3ec3def4062fe461803ad88bce9c8

Former-commit-id: a0dc31f3c53cde9cc5e23d8b6e3a6c2af63fbcf2