Do not allow editing of mcc, mnc for carrier id APNs
A user that is allowed to edit mcc, mnc or mvno data can break
the APN list. If a carrier or OEM has added an APN that is
matched with carrier id there is no need to edit those fields.

Apn Editor should not throw any error when a user backs out of
from a carrier id APN where mcc, mnc is empty.

Disable editing of these fields and check for carrier id APN
before adding error message.

Test: Add carrier APN, insert sim and back out from ApnEditor
Bug: 222103038
Change-Id: I4e86cdd3e8caaf3dbfefe9b9f82f11c2804577ff

Former-commit-id: aa4ac1e5b98f3b5f3d38ef13f5dc286a4a0fb642