Fix LaunchAnyWhere in AppRestrictionsFragment
If the intent's package equals to the app's package, this intent
will be allowed to startActivityForResult.
But this check is unsafe, because if the component of this intent
is set, the package field will just be ignored. So if we set the
component to any activity we like and set package to the app's
package, it will pass the assertSafeToStartCustomActivity check
and now we can launch anywhere.

Bug: 223578534
Test: robotest and manual verify
Change-Id: I40496105bae313fe5cff2a36dfe329c1e2b5bbe4
(cherry picked from commit e60d5c88fa949d77ea7612d7b383d86772af9c90 [formerly 90e095dbe372f29823ad4788c0cc2d781ae3bb24])

Former-commit-id: f8f45888e6d20b238b222b95d18898fa1ab81ed4