Avoid accessibility pages display on notification bar in setup
Root cause: ag/7505195 introduce to set clipToPadding false on scrollView to make it extend below navbar to overlaps with sibling views. In this case anything pinned on top of screen (switchbar, pinned header for most UIs, and preview indicator for font size page).So, the recycler draw content outside of its bounds.
Solution: Defines whether a child is limited to draw inside of its bounds and the ViewGroup will clip its drawing surface so as to exclude the padding area for SuW flow.

Bug: 227349626
Bug: 227409455
Bug: 222022439
Test: Manual testing
Change-Id: I439508c8e8c90e7303a72a4990c7ef1b9e5bc19d

Former-commit-id: 7280add0321f02d511712988d0be8480c6419d0a