MainClear: remove obsolete check for multi-volume FDE
It is never the case that '!isExternalStorageEmulated() &&
!isExternalStorageRemovable() && isExtStorageEncrypted()', so stop
checking for this.

It's easy to see that even just the last part, isExtStorageEncrypted(),
is never true, since it checks whether the property "vold.decrypt" is
non-empty.  That system property was only set on devices that used FDE
(Full Disk Encryption), and FDE support was removed in Android T.

Moreover, the full condition of '!emulated && !removable && FDE'
actually checked for devices that used multi-volume FDE, which
apparently was never actually used, and was removed 8 years ago by vold
commit 87999173dd79 (http://ag/423222).  This is consistent with the
fact that this code was added 11 years ago by commit 2a32192329ac
(http://ag/134210), when probably this feature was still planned.

Bug: 208476087
Change-Id: I4c2b5c9aed8812367a70d0dd1bbcc94008fec529

Former-commit-id: 8c25415736466323f8b2493845c320155bcc471b