[ALPSxxxxxxxx] Storage view is white and cannot show any info
Storage view is white and cannot show any info while plug in/out u disk because the visibility of list view is gone.
This visibility is set when StorageDashboardFragment.java sets loading using API setLoading.
In error condition setloading was not called due to conditional statement.

After removing the conditional statement, it sets loading through which visibility of the list view is set to VISIBLE.

1. Enter Settings->Storage
2. Plug in U-disk, storage view will start updating
3. During storage view updating, plug out U-disk(before phone recognized and show U-disk)

Fix:
Bug:
Test:
Change-Id: Id7350909b614bbe9264910cab3cdeca616544618
CR-Id:
Feature:

Former-commit-id: 800152033e198a9b71dde485af6a4d4047dec72e