Make sure loading extra info once
In those pages that extend ManageApplications, they need to load extra
information from other services as a summary for each package or as a
condition to be filtered. The way to load extra information is invoked
at two places. One is invoked once all apps are loaded and the other is
running before all apps loaded. It took much more time to prepare the
data for those pages if it is time consuming to load extra info. This
also caused the loading spinner to be always shown.

This change is to make sure that the loading extra info is invoked just
once when entering those pages and to decrease the time to load extra
info. The loading spinner is still displayed on some pages since it
still needs to take more than 100 ms to load extra infor in those pages.

Bug: 222441965
Bug: 223315960
Bug: 222981538
Bug: 222981533
Test: manual test
Change-Id: I8d0084ab5c412d9661f31446e613b7d1f58202ca

Former-commit-id: b8315994b9dcc7e322c69498e51ed259418000b5