Add FLAG_ACTIVITY_SINGLE_TOP to shortcuts intent
This is to simplify the shortcut launch flow on large screen devices
when launching the same shortcut, which also improves the performance.

FLAG_ACTIVITY_CLEAR_TOP flag is used for the launch a shortcut. When
launching a shortcut whose activity is in the background on a large
screen device, the background activity is cleared by the system and a
new activity is started. Thus, the activity goes though the complete
deep link flow that invokes the trampoline homepage activity to start
the target page.

After the change, the system will bring the background activity to the
foreground rather than kill it and start a new one.

Test: create a shortcut and observe the adb log
Fix: 234681140
Change-Id: Id922af1f59a6a64ea6f6150cf6fea92808a6de65

Former-commit-id: d3286b198124fa21b86fdd99b68fb0fc44e0ec37