Bool to note Pending Intent bug fixed
A workaround can be used to cover for the fact that PendingIntents sent
by Settings might differ only by Intent extras (and thus be wrong). This
bool notes that the issue is fixed so the workaround code can be
dropped.

Bug: 238605613
Test: manually
Merged-In: I478dd5fbd53d1dafd049e392a186d9796dd179cb
Change-Id: Ife355166df080ce7371a95b5cea28575227153f0

Former-commit-id: f952e89d4bf09343c377aa51fe23281ff82a1069